Move expected methods in ExpectModelUtils
